# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Simple AI Gateway",
        "description": "LLM API \u7f51\u5173 \u2014 \u517c\u5bb9 OpenAI API \u534f\u8bae\uff0c\u652f\u6301\u591a\u4f9b\u5e94\u5546\u8def\u7531\u3001\u7194\u65ad\u3001\u9650\u6d41\u3001\u914d\u989d\u3001\u52a8\u6001\u914d\u7f6e\u7ba1\u7406\u4e0e Webhook \u4e8b\u4ef6\u901a\u77e5\u3002\n\n## API \u5206\u7ec4\n- **\u4e1a\u52a1\u7aef\u70b9**: `/v1/chat/completions`, `/v1/models`\n- **\u8ba4\u8bc1**: `/auth/*`\n- **\u7ba1\u7406**: `/admin/*`\n- **\u5185\u90e8**: `/internal/*`\n- **\u7cfb\u7edf**: `/healthz`\n\n## \u8ba4\u8bc1\u65b9\u5f0f\nBearer token: JWT (access token) \u6216 API Key",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        },
        {
            "ApiKeyAuth": []
        }
    ],
...
    "paths": {
        "/auth/password": {
            "put": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "changePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "summary": "Refresh an expired access token using a valid refresh token",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "Invalid or expired refresh token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GatewayErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "New access token issued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/internal/providers/runtime": {
            "get": {
                "tags": [
                    "internal-provider-state-controller"
                ],
                "operationId": "runtime",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderRuntimeSnapshotResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/internal/providers/discovery": {
            "get": {
                "tags": [
                    "internal-provider-state-controller"
                ],
                "operationId": "discovery",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscoverySnapshot"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/providers": {
            "get": {
                "tags": [
                    "admin-provider-controller"
                ],
                "operationId": "providers",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProvidersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChangePasswordRequest": {
                "required": [
                    "newPassword",
                    "oldPassword"
                ],
                "type": "object",
                "properties": {
                    "oldPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                }
            },
...
            "GatewayErrorResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "invalid_request"
                    },
                    "message": {
                        "type": "string",
                        "example": "model is required"
                    },
                    "requestId": {
                        "type": "string",
                        "example": "req_01HXYZ"
                    }
                },
                "description": "Normalized gateway error response"
            },
...
            "RefreshRequest": {
                "required": [
                    "refreshToken"
                ],
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "description": "Refresh token",
                        "example": "eyJhbGciOiJI..."
                    }
                },
                "description": "JWT refresh request payload"
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "description": "Access token",
                        "example": "eyJhbGciOiJI..."
                    },
                    "refreshToken": {
                        "type": "string",
                        "description": "Refresh token",
                        "example": "eyJhbGciOiJI..."
                    },
                    "tokenType": {
                        "type": "string",
                        "description": "Token type",
                        "example": "Bearer"
                    }
                },
                "description": "JWT login/refresh response"
            },
...
            "ProviderRuntimeSnapshotResponse": {
                "type": "object",
                "properties": {
                    "generatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "providers": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ProviderRuntimeState"
                        }
                    }
                }
            },
            "ProviderRuntimeState": {
                "type": "object",
                "properties": {
                    "runtimeAvailable": {
                        "type": "boolean"
                    },
                    "lastCheckedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastSuccessAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "consecutiveFailures": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "consecutiveSuccesses": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "httpStatus": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "latencyMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "reason": {
                        "type": "string"
                    }
                }
            },
            "DiscoverySnapshot": {
                "type": "object",
                "properties": {
                    "providers": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ProviderDiscovery"
                        }
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ProviderDiscovery": {
                "type": "object",
                "properties": {
                    "source": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "lastFetchedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "models": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "error": {
                        "type": "string"
                    }
                }
            },
...
            "ProviderView": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "baseUrl": {
                        "type": "string"
                    },
                    "apiKey": {
                        "type": "string"
                    },
                    "keys": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "keyWeights": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "timeout": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "models": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ProvidersResponse": {
                "type": "object",
                "properties": {
                    "generatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "providers": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ProviderView"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT access token in Authorization header: Bearer <token>",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Gateway client API key in Authorization header",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}